Merging staging into master

+++ server/api/v1.1/custom_fields/custom_field.controller.js
@@ -4,12 +4,12 @@ const callApi = require('../utility')
 const { sendSuccessResponse, sendErrorResponse } = require('../../global/response')
 
 exports.index = function (req, res) {
-  callApi.callApi('companyuser/query', 'post', { domain_email: req.user.domain_email })
+  callApi.callApi('companyuser/query', 'post', { domain_email: req.user.domain_email }, req.headers.authorization)
     .then(companyUser => {
       if (!companyUser) {
         sendErrorResponse(res, 404, '', 'The user account does not belong to any company. Please contact support')
       }
-      callApi.callApi('custom_fields/query', 'post', { purpose: 'findAll', match: { companyId: companyUser.companyId } })
+      callApi.callApi('custom_fields/query', 'post', { purpose: 'findAll', match: { companyId: companyUser.companyId } }, req.headers.authorization)
         .then(customFields => {
           sendSuccessResponse(res, 200, customFields)
         })
@@ -27,7 +27,7 @@ exports.index = function (req, res) {
 }
 
 exports.create = function (req, res) {
-  callApi.callApi('companyUser/query', 'post', { domain_email: req.user.domain_email })
+  callApi.callApi('companyUser/query', 'post', { domain_email: req.user.domain_email }, req.headers.authorization)
     .then(companyUser => {
       if (!companyUser) {
         sendErrorResponse(res, 404, '', 'The user account does not belong to any company. Please contact support')
@@ -39,7 +39,7 @@ exports.create = function (req, res) {
         companyId: companyUser.companyId,
         createdBy: req.user._id
       }
-      callApi.callApi('custom_fields/', 'post', customFieldPayload)
+      callApi.callApi('custom_fields/', 'post', customFieldPayload, req.headers.authorization)
         .then(newCustomField => {
           require('./../../../config/socketio').sendMessageToClient({
             room_id: companyUser.companyId,
@@ -62,7 +62,7 @@ exports.create = function (req, res) {
 }
 
 exports.update = function (req, res) {
-  callApi.callApi('custom_fields/query', 'post', { purpose: 'findOne', match: { _id: req.body.customFieldId, companyId: req.user.companyId } })
+  callApi.callApi('custom_fields/query', 'post', { purpose: 'findOne', match: { _id: req.body.customFieldId, companyId: req.user.companyId } }, req.headers.authorization)
     .then(fieldPayload => {
       if (!fieldPayload) {
         sendErrorResponse(res, 404, '', 'No Custom field is available on server with given customFieldId.')
@@ -72,7 +72,7 @@ exports.update = function (req, res) {
       if (req.body.updated.type) updatedPayload.type = req.body.updated.type
       if (req.body.updated.description) updatedPayload.description = req.body.updated.description
       if (req.user.companyId) updatedPayload.companyId = req.user.companyId
-      callApi.callApi('custom_fields/', 'put', { purpose: 'updateOne', match: { _id: req.body.customFieldId }, updated: updatedPayload })
+      callApi.callApi('custom_fields/', 'put', { purpose: 'updateOne', match: { _id: req.body.customFieldId }, updated: updatedPayload }, req.headers.authorization)
         .then(updated => {
           require('./../../../config/socketio').sendMessageToClient({
             room_id: fieldPayload.companyId._id,
@@ -95,12 +95,12 @@ exports.update = function (req, res) {
 }
 
 exports.delete = function (req, res) {
-  callApi.callApi('custom_field_subscribers/query', 'post', { purpose: 'findOne', match: { customFieldId: req.body.customFieldId } })
+  callApi.callApi('custom_field_subscribers/query', 'post', { purpose: 'findOne', match: { customFieldId: req.body.customFieldId } }, req.headers.authorization)
     .then(foundCustomField => {
       if (foundCustomField) {
-        callApi.callApi('custom_field_subscribers/', 'delete', { purpose: 'deleteMany', match: { customFieldId: req.body.customFieldId } })
+        callApi.callApi('custom_field_subscribers/', 'delete', { purpose: 'deleteMany', match: { customFieldId: req.body.customFieldId } }, req.headers.authorization)
           .then(() => {
-            callApi.callApi('custom_fields/', 'delete', { purpose: 'deleteOne', match: { _id: req.body.customFieldId } })
+            callApi.callApi('custom_fields/', 'delete', { purpose: 'deleteOne', match: { _id: req.body.customFieldId } }, req.headers.authorization)
               .then(fieldPayload => {
                 require('./../../../config/socketio').sendMessageToClient({
                   room_id: fieldPayload.companyId,
@@ -121,7 +121,7 @@ exports.delete = function (req, res) {
             sendErrorResponse(res, 500, `Failed to remove custom field subscriber${err}`)
           })
       } else {
-        callApi.callApi('custom_fields/', 'delete', { purpose: 'deleteOne', match: { _id: req.body.customFieldId } })
+        callApi.callApi('custom_fields/', 'delete', { purpose: 'deleteOne', match: { _id: req.body.customFieldId } }, req.headers.authorization)
           .then(fieldPayload => {
             require('./../../../config/socketio').sendMessageToClient({
               room_id: fieldPayload.companyId,
